+++ POMPOM/POMPOM/UI/ClothPickerView/ViewModel/PickerViewModel.swift
@@ -60,35 +60,3 @@ class PickerViewModel: ObservableObject {
     }
 }
 
-struct Cloth: Identifiable {
-    var id: Int
-    var name: String
-    var category: ClothCategory
-}
-
-enum ClothCategory: CaseIterable, Identifiable {
-    var id: ClothCategory {
-        self
-    }
-    
-    case hat
-    case top
-    case bottom
-    case socks
-    case shoes
-    
-    var koreanSubtitle: String {
-        switch self {
-        case .hat:
-            return "모자"
-        case .top:
-            return "상의"
-        case .bottom:
-            return "하의"
-        case .socks:
-            return "양말"
-        case .shoes:
-            return "신발"
-        }
-    }
-}
